Return the contents of a config file

Return the contents of a config file

read_config_file(conf = NA)

Arguments

conf

String. Either the path and name of a config file, or 'site', or 'user', or NA (the default). If NA, get_config will search the current directory for a config file.

Value

optionally nested named list or vector as returned by yaml::read_yaml, or NA if the appropriate file does not exist.

Details

The default search name for config files is "sqlhelper_db_conf.yml". If conf is 'site', 'user' or NA, a file with this name will be sought in rappdirs::site_config_dir, rappdirs::user_config_dir or the current working directory, respectively. In an interactive context, a warning will be issued if the sought file does not exist.