Execute a sequence of SQL queries

Accepts a character vector of SQL queries and runs each one

Usage

runqueries(
  sql,
  quotesql = TRUE,
  values = parent.frame(),
  execmethod = "get",
  geometry = NA,
  default_conn = live_connection(get_default_conn_name()),
  include_params = FALSE
)

Arguments

sql

An optionally-named list or character vector containing sql commands, or a tibble returned by read_sql()

quotesql

"yes" or "no" - should interpolated characters be quoted by default?

execmethod

One of "get", "execute", "sendq", "sends" or "spatial" - which method should be used to execute the query? "get" means DBI::dbGetQuery(); "execute" means DBI::dbExecute(); "sendq" means DBI::dbSendQuery; "sends" means DBI::dbSendStatement(); "spatial" means sf::st_read().

geometry

If execmethod is "spatial", which geometry column should be used (ignored if execmethod is not spatial)

include_params

Should the parameters be included in the output?

interpolate

Should the SQL be parameterized from R? Defaults to the value of parent.frame(). Pass any object that is not an environment (e.g. "no" or FALSE) if interpolation is to be skipped, or another environment containing values to interpolate to avoid using .GlobalEnv.

conn

Either the name of a sqlhelper connection, or a database connection returned by DBI::dbConnect(), or NA

Value

If include_params is FALSE, an named list containing the results of each query; the names are the same as those in the sql

parameter. If include_params is TRUE, a tibble containing the query as executed, the above parameters, and the result of each query.

See also